refactor(domain): 将 BaseRepository 的接口方法统一风格
